+++ service/__init__.py
@@ -23,6 +23,7 @@ from flask import Flask
 from flask_restx import Api
 from service import config
 from service.common import log_handlers
+from service.models import models
 
 
 # Will be initialize when app is created
@@ -52,14 +53,14 @@ def create_app():
         prefix="/api",
     )
 
-    from service.models import db
+    from service.models.models import db
 
     db.init_app(app)
 
     with app.app_context():
         # Dependencies require we import the routes AFTER the Flask app is created
         # pylint: disable=wrong-import-position, wrong-import-order, unused-import
-        from service import routes, models  # noqa: F401 E402
+        from service.routes import routes  # noqa: F401 E402
         from service.common import error_handlers, cli_commands  # noqa: F401, E402
 
         try:
