+++ service/common/cli_commands.py
@@ -17,7 +17,7 @@
 Flask CLI Command Extensions
 """
 from flask import current_app as app  # Import Flask application
-from service.models import db
+from service.models.models import db
 
 
 ######################################################################
