+++ tests/factories.py
@@ -4,7 +4,7 @@ Test Factory to make fake objects for testing
 
 import factory
 from factory.fuzzy import FuzzyChoice
-from service.models import Product, Status
+from service.models.models import Product, Status
 
 
 class ProductFactory(factory.Factory):
